+++ python/jaeger/fps.py
@@ -240,6 +240,8 @@ class FPS(BaseFPS["FPS"]):
 
         self.observatory = config["observatory"]
 
+        self.disabled: set[int] = set([])
+
         if self.ieb is None or self.ieb is True:
             self.ieb = config["files"]["ieb_config"]
 
@@ -392,6 +394,7 @@ class FPS(BaseFPS["FPS"]):
         self: T,
         start_pollers: bool | None = None,
         enable_low_temperature: bool = True,
+        keep_disabled: bool = True,
     ) -> T:
         """Initialises all positioners with status and firmware version.
 
@@ -399,12 +402,23 @@ class FPS(BaseFPS["FPS"]):
         ----------
         start_pollers
             Whether to initialise the pollers.
+        enable_low_temperature
+            Enables the low temperature warnings.
+        keep_disabled
+            Maintain the list of disabled/offline robots.
 
         """
 
         if start_pollers is None:
             start_pollers = config["fps"]["start_pollers"]
         assert isinstance(start_pollers, bool)
+
+        if keep_disabled:
+            for positioner in self.positioners.values():
+                if positioner.offline or positioner.disabled:
+                    self.disabled.add(positioner.positioner_id)
+        else:
+            self.disabled = set([])
 
         # Clear all robots
         self.clear()
@@ -466,8 +480,12 @@ class FPS(BaseFPS["FPS"]):
             positioner.fps = self
             positioner.firmware = get_fw_command.get_firmware()[reply.positioner_id]
 
-            if positioner.positioner_id in config["fps"]["disabled_positioners"]:
+            if (
+                positioner.positioner_id in config["fps"]["disabled_positioners"]
+                or positioner.positioner_id in self.disabled
+            ):
                 positioner.disabled = True
+                self.disabled.add(positioner.positioner_id)
 
         # Add offline robots. Offline positioners are physically in the array but
         # they don't reply to commands and we need to specify their position. Once
@@ -483,6 +501,8 @@ class FPS(BaseFPS["FPS"]):
                 positioner.offline = True
                 positioner.alpha = off_alpha
                 positioner.beta = off_beta
+
+                self.disabled.add(positioner.positioner_id)
 
         # Mark as initialised here although we have some more work to do.
         self.initialised = True